Accept an `after` value for repository pagination.
This introduces an `after` parameter that can be supplied to a `repositories` query for pagination purposes:

```
{
  repositories(first: 1, after: "UmVwb3...") {
    nodes {
      name
      description
      url
    }
    pageInfo {
      endCursor
    }
  }
}
```

Note that from the client's perspective, the GraphQL parameter is a string (specifically, a Base64-encoded opaque value). An intermediate transformation will handle unmarshalling the string into a dedicated cursor type and taking an appropriate action.